pk error

+++ StoreConsoleApp/StoreConsoleApp/Entities/Product.cs
@@ -11,7 +11,7 @@ namespace ConsoleApp1
         public int ProductId { get; set; }
         public string ProductName { get; set; }
         
-        public int ModelYear { get; set; }
+        public short ModelYear { get; set; }
         public decimal ListPrice { get; set; }
 
         // one to many relationship between Products and Brands
